trying out 16bit

+++ cordic_V_fixed_point.c
@@ -1,15 +1,15 @@
 
 int z_table[15];
 
-void cordic_V_fixed_point(int *x, int *y, int *z)
+void cordic_V_fixed_point(short *x, short *y, short *z)
 {
-    int x_temp_1, y_temp_1, z_temp;
-    int x_temp_2, y_temp_2;
-    int i;
+    short x_temp_1, y_temp_1, z_temp;
+    short x_temp_2, y_temp_2;
+    short i;
     x_temp_1 = *x;
     y_temp_1 = *y;
     z_temp = 0;
-    for (int i = 0; i < 15; i++)
+    for (short i = 0; i < 15; i++)
     { /* 15 iterations are needed */
         if (y_temp_1 > 0)
         {
